Dealing with promiscuous decorators.
1) the marker handler does not add messages to the message panel anymore.  That
was my mistake.

2) the EmfFormEditor has a method that returns active markers.

2) I changed EmfDetailsPart to retain the Bindings created by derived classes.

3) a method in EmfDetailsPart, addMessagesFromMarkers, compares the retained
Bindings to the editor's active Markers.  From the Marker, it gets an EObject
and finds it as a model in one of the retained bindings.  It adds a message to
the message panel with the swt control from the binding and the message string
from the Marker and using the EObject as key.

4) addMessagesFromMarkers is invoked by EmfDetailsPart.selectionChanged method.
 This replaces the EmfFormEditor.validate call and may run faster than a full
validation.

5) The selectionChanged handler for the viewer in EmfMasterDetailDetailBlock
deletes the messages from the message manager.  It appears this is called by
the eclipse ui before the call to EmfDetailsPart.selectionChanged.

So, the end of all this is that when any builder creates a marker for an
EObject that happens to be bound in a details part, the marker's message will
be displayed on the top panel of the editor (as well as in the Problem's view),
and a decorator will appear on the bound control.

Basically, the Problems View is synchronized with the message panel.

It doesn't look so smooth in the ds example as it does in my model editor.
Some differences there in assumptions.  I haven't pursued.